将读取的图片存储到硬盘作为下次的缓存，增快出图速度
